시은 - 실시간 환경데이터 db 연동

+++ SpringMVC03/src/main/java/kr/board/entity/EnvCri.java
@@ -14,7 +14,7 @@ public class EnvCri {
 	private float temperature;
 	private float humidity;
 	private int co2;
-	private int ammonia;
+	private float ammonia;
 	private float pm;
 	private String created_at;
